Change to latest NordVPN 3.14.1
Fixed: Some users had an issue with the app freezing at random. Rumor has it, some were even confronted with the never-ending spinner. The bug causing the trouble is fixed now, and the app is set to run smoothly without awkward silent interruptions.
Better: When opened, the app will now greet you with a list of useful commands displayed in the terminal

NordLynx is now the default VPN protocol in our Linux app.
We made changes in firewall handling. Now we filter packets by firewall marks instead of IP addresses of VPN servers.
A few performance improvements are in place.

We switched from destination-based to policy-based traffic routing.
Fixed: A bug that slipped into the previous release and prevented OpenVPN from working on any devices except ones with the AMD64 architecture.